requirement.txt
aiohttp==3.8.5
aiosignal==1.3.1
altgraph==0.17.3
ansi2html==1.8.0
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
auto-py-to-exe==2.36.0
beautifulsoup4==4.12.2
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.7
colour==0.1.5
cycler==0.11.0
dash==2.14.1
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
docopt==0.6.2
Eel==0.16.0
et-xmlfile==1.1.0
fastjsonschema==2.18.0
Flask==2.2.5
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.3
getfilelistpy==1.0.7
gevent==22.10.2
gevent-websocket==0.10.1
google==3.0.0
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-cloud-vision==3.4.4
google-services-api==1.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.2
grpcio-status==1.56.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
jupyter-core==4.12.0
kiwisolver==1.4.4
macholib==1.16.2
MarkupSafe==2.1.3
matplotlib==3.5.3
multidict==6.0.4
nbformat==5.8.0
nest-asyncio==1.5.8
numpy==1.21.6
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==1.1.5
patsy==0.5.3
pigar==2.1.1
Pillow==9.5.0
pipreqs==0.4.13
pkgutil-resolve-name==1.3.10
plotly==5.18.0
proto-plus==1.22.3
protobuf==4.23.4
psycopg2-binary==2.9.6
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyDrive==1.3.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
pyparsing==3.1.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
retry==0.9.2
retrying==1.3.4
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0.post7
soupsieve==2.4.1
statsmodels==0.13.5
tenacity==8.2.3
threadpoolctl==3.1.0
tkmacosx==1.0.5
traitlets==5.9.0
typing-extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.6
Werkzeug==2.2.3
whichcraft==0.6.1
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
